file check while deleted

+++ app/Nrna/Services/PostService.php
@@ -199,7 +199,10 @@ class PostService
                 $formData = $this->getVideoData($formData);
             }
             if (isset($formData['metadata']['featured_image'])) {
-                $this->file->delete($this->uploadPath . '/' . $post->metadata->featured_image);
+                if(isset($post->metadata->featured_image)){
+                    $this->file->delete($this->uploadPath . '/' . $post->metadata->featured_image);
+                }
+
                 $featuredInfo                           = $this->fileUpload->handle(
                     $formData['metadata']['featured_image'],
                     $this->uploadPath
